Add syntax highlighiting to code output
