+++ Cargo.toml
@@ -16,6 +16,6 @@ persistence = ["eframe/persistence", "egui_extras/serde"]
 chrono = "0.4.42"
 eframe = "0.33"
 egui = { version = "0.33", features = ["serde"] }
-egui_extras = { version = "0.33", features = ["chrono"] }
+egui_extras = { version = "0.33", features = ["chrono", "syntect"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
